+++ frontend/products/products_dashboard.js
@@ -61,6 +61,7 @@ function fetchProductsList() {
 
 function renderTable(products) {
     let productsTable = document.getElementById("table-body");
+    const accessToken = localStorage.getItem("access");
     productsTable.innerHTML = "";
 
     products.forEach(product => {
@@ -101,26 +102,44 @@ function renderTable(products) {
         deleteButton.style.backgroundColor = 'red';
 
         deleteButton.addEventListener('click', () => {
-            fetch(`http://127.0.0.1:8000/api/products/${product.id}/`, {
-                method: "DELETE",
-                headers: { 
-                    "Authorization": `Bearer ${accessToken}`,
+            Swal.fire({
+                title: "Are you sure?",
+                text: `Do you really want to delete ${product.name}?`,
+                icon: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#dc3545",
+                cancelButtonColor: "#6c757d",
+                confirmButtonText: "Yes, delete it!"
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    fetch(`http://127.0.0.1:8000/api/products/${product.id}`, {
+                        method: "DELETE",
+                        headers: { 
+                            "Authorization": `Bearer ${accessToken}`,
+                            "Content-Type": "application/json"
+                        }
+                    })
+                    .then(response => {
+                        if (!response.ok) throw new Error("Deletion failed");
+        
+                        Swal.fire("Deleted!", "The employee has been removed.", "success");
+                        fetchEmployeeList(); // Tabelle aktualisieren
+                    })
+                    .catch(error => {
+                        console.error("Error:", error);
+                        Swal.fire("Error", "Delete failed!", "error");
+                    });
                 }
-            })
-            .then(response => {
-                if (!response.ok) throw new Error(`Deletion error: ${response.status} ${response.statusText}`);
-                    console.log(`Products with ID ${product.id} was deleted.`);
-                    return response.text();
-            })
-            .then(() => fetchProductsList())
-            .catch(error => {
-                console.error("Error:", error);
-                alert(`Delete failed! Error: ${error.message}`);
             });
         });
-
+        
         deleteButtonCell.appendChild(deleteButton);
         row.appendChild(deleteButtonCell);
+
         productsTable.appendChild(row);
     });
 }
+
+function add(){
+    window.location.href= '../edit/edit.html';
+  }
